requirements.txt
Django==1.5
MySQL-python==1.2.4
PIL==1.1.7
South==0.8.4
argparse==1.2.1
django-classy-tags==0.4
django-cms==2.4.3
django-modeltranslation==0.7.2
django-mptt==0.5.2
django-rosetta==0.7.2
django-sekizai==0.7
django-tinymce==1.5.2
html5lib==1.0b3
six==1.4.1
wsgiref==0.1.2
